+++ packages/afformative/src/index.js
@@ -1,3 +1,2 @@
 export * from "./constants"
 export { default as makeFormatter } from "./makeFormatter"
-export { default as makeUseFormatter } from "./makeUseFormatter"