+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 0.3.0 [2020-06-24]
 
-Nothing!
+### Added
+
+- `FORMATTER_OVERRIDE` to allow overriding the behaviour of all formatters for certain values.
+
+### Removed
+
+- `makeUseFormatter` has been removed in favour of manual hook creation.
+
+### Changed
+
+- `Formatter.format` now adds `SUGGESTIONS.primitive` only if no suggestions are passed. This allows using `Formatter.format` as a way to delegate formatting without any additional logic.
 
 ## 0.2.0 [2020-05-14]